+++ cmd/dist.js
@@ -18,36 +18,22 @@ if (!moduleName) {
 
 checkConfigFile();
 
-const projectConfig = require('../project_config');
+// Guarantee `share.originalProcessArgv` has been loaded.
+const share = require('../share');
+const pathInfo = require('../data/path_info');
 
-require('./util/lint')(() => {
-  // Guarantee `share.originalProcessArgv` has been loaded.
-  const share = require('../share');
-  const pathInfo = require('../data/path_info');
+// Modify `process.argv` for `gulp-cli`.
+process.argv = [share.originalProcessArgv[0], share.originalProcessArgv[1], 'dist', '--gulpfile', pathInfo.gulpFile];
 
-  if (projectConfig.onlyWebpack) {
-    require('./util/webpack');
-  } else {
-    // Modify `process.argv` for `gulp-cli`.
-    process.argv = [
-      share.originalProcessArgv[0],
-      share.originalProcessArgv[1],
-      'dist',
-      '--gulpfile',
-      pathInfo.gulpFile,
-    ];
-
-    require('gulp-cli')(err => {
-      if (err) {
-        logger.error(`
+require('gulp-cli')(err => {
+  if (err) {
+    logger.error(`
   Error occurred when lila build modules, you should resolve those errors, and try again.
-        `);
-        logger.error(err.stack || err);
-      } else {
-        logger.success(`
+    `);
+    logger.error(err.stack || err);
+  } else {
+    logger.success(`
   Pack source codes and static files into production successfully.
-        `);
-      }
-    });
+    `);
   }
 });