+++ examples/dev/eslint.js
@@ -1,4 +1,0 @@
-
-require('../../util/change_cwd')(__dirname + '/demo');
-
-require('../../util/exec')('lila eslint test-6/index');