+++ examples/dev/prettier.js
@@ -1,4 +1,0 @@
-
-require('../../util/change_cwd')(__dirname + '/demo');
-
-require('../../util/exec')('lila prettier test-6/index,test-5/*');