+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.4.1 [2018-08-01]
+
+- Remove `upgrade` command.
+- Remove `onlyWebpack` option.
+- Remove `esLintOptions, esLintFix, styleLintOptions, styleLintFix, prettierOptions` option.
+- Remove `esLint, styleLint, prettier` option.
+- Remove `eslint,stylelint,prettier` command.
+- Update `project_files/root/_package.json`.
+
 ## 0.4.0 [2018-07-26]
 
 - Remove `archive,arc,clean,revert` all server-side commands.